Rename and move logprep.output.es_output to logprep.connector.elasticsearch.output
