+++ tests/unit/connector/test_connector_factory.py
@@ -26,7 +26,7 @@ from logprep.connector.confluent_kafka.output import (
     ConfluentKafkaOutput,
 )
 from logprep.connector.dummy.output import DummyOutput
-from logprep.output.es_output import ElasticsearchOutput
+from logprep.connector.elasticsearch.output import ElasticsearchOutput
 from logprep.output.writing_output import WritingOutput
