+++ tests/unit/connector/test_elasticsearch_output_factory.py
@@ -11,7 +11,7 @@ from pytest import fail, raises
 from logprep.connector.confluent_kafka.input import (
     ConfluentKafkaInputFactory,
 )
-from logprep.output.es_output import ElasticsearchOutputFactory
+from logprep.connector.elasticsearch.output import ElasticsearchOutputFactory
 from logprep.connector.connector_factory import InvalidConfigurationError
